+++ include/d/actor/d_a_tag_island.h
@@ -31,7 +31,7 @@ public:
     BOOL otherCheck();
     BOOL arrivalTerms();
     void demoInitProc();
-    int next_msgStatus(unsigned long*);
+    int next_msgStatus(u32*);
     u32 getMsg();
     void talkInit();
     u16 talk();
